Move components to src/components
