+++ src/App.js
@@ -1,23 +1,23 @@
 import './App.css';
-import Header from './Header';
-import Splash from './Splash';
-import About from './About';
-import Events from './Events'
-import Members from './Members';
-import Footer from './Footer'
+import Header from './components/Header';
+import Splash from './components/Splash';
+import About from './components/About';
+import Events from './components/Events'
+import Members from './components/Members';
+import Footer from './components/Footer'
 
 function App() {
-  return (
-    <div className="App">
-      <Header />
-      <Splash />
-      <About />
-      <Events />
-      <img src='/divider.jpg' alt='code on a screen. mostly css' />
-      <Members />
-      <Footer />
-    </div>
-  );
+    return (
+        <div className="App">
+            <Header />
+            <Splash />
+            <About />
+            <Events />
+            <img src='/divider.jpg' alt='code on a screen. mostly css' />
+            <Members />
+            <Footer />
+        </div>
+    );
 }
 
 export default App;
