+++ src/components/Splash.jsx
