+++ src/test/components/About.test.js
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import About from "../../About";
+import About from "../../components/About";
 
 describe("About", () => {
     it("renders", () => {
