+++ src/test/components/Events.test.js
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import Events from "../../Events";
+import Events from "../../components/Events";
 
 describe("Events", () => {
     it("renders", () => {
