+++ src/test/components/Header.test.js
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import Header from "../../Header";
+import Header from "../../components/Header";
 
 describe("Header", () => {
     it("renders", () => {
